move stream handlers into concrete stream
